web_crawler/car_finder/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
rich>=13.0.0
tqdm>=4.65.0
selenium>=4.15.0
flask>=2.3.0
gunicorn>=21.0.0 